requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
anthropic==0.40.0
pydantic==2.9.0
pytest==8.3.0
httpx==0.27.0
